+++ app/preload.js
@@ -1,7 +1,7 @@
 const { contextBridge, ipcRenderer } = require("electron");
 
-function invoke(channel, payload) {
-	return ipcRenderer.invoke(channel, payload);
+function invoke(...args) {
+	return ipcRenderer.invoke(...args);
 }
 
 contextBridge.exposeInMainWorld("ebApi", {